backend/requirements.txt
autopep8
flake8
fastapi
uvicorn[standard]
google-cloud-speech
